requirements.txt
discord.py==2.3.2
APScheduler==3.10.4
python-dotenv==1.0.0
Pillow==10.0.0
aiohttp==3.9.1
openai==1.12.0
